+++ packages/pluggableWidgets/gallery-web/src/Gallery.tsx
@@ -12,7 +12,12 @@ import {
 import { FilterCondition } from "mendix/filters";
 import { extractFilters } from "./utils/filters";
 import { and } from "mendix/filters/builders";
-import { executeAction, getGlobalSelectionContext, useSelectionHelper } from "@mendix/pluggable-widgets-commons";
+import {
+    executeAction,
+    getGlobalSelectionContext,
+    useCreateSelectionContextValue,
+    useSelectionHelper
+} from "@mendix/pluggable-widgets-commons";
 
 export function Gallery(props: GalleryContainerProps): ReactElement {
     const viewStateFilters = useRef<FilterCondition | undefined>(undefined);
@@ -104,24 +109,7 @@ export function Gallery(props: GalleryContainerProps): ReactElement {
 
     const selection = useSelectionHelper(props.itemSelection, props.datasource, props.onSelectionChange);
 
-    const toggleSelection = useCallback(() => {
-        if (selection?.type === "Multi") {
-            if (selection.selectionStatus === "all") {
-                selection.selectNone();
-            } else {
-                selection.selectAll();
-            }
-        }
-    }, [selection]);
-    const multiSelectionStatus = selection?.type === "Multi" ? selection.selectionStatus : undefined;
-    const selectionContextValue = useMemo(() => {
-        if (multiSelectionStatus !== undefined) {
-            return {
-                status: multiSelectionStatus,
-                toggle: toggleSelection
-            };
-        }
-    }, [multiSelectionStatus, toggleSelection]);
+    const selectionContextValue = useCreateSelectionContextValue(selection);
 
     const showHeader = props.filterList.length > 0 || props.sortList.length > 0 || selection?.type === "Multi";